refactor: CartProvider에서 외부의 훅을 불러오는 것이 아닌 컴포넌트 내에서 호출하도록 수정
