+++ src/contexts/CartProvider.tsx
@@ -1,9 +1,75 @@
 import React from "react";
-import { CartContext } from "./CartContext";
-import useShoppingCart from "../hooks/useShoppingCart";
+import { useCallback, useEffect, useState, useMemo } from "react";
+import { CartItemTypes } from "../types/CartItemType";
+import getShoppingCart from "../api/getShoppingCart";
+import { findIsCartItem, isCartFull } from "../utils/cart";
+import { CartContext } from "./useCartContext";
+
+type Status = "idle" | "loading" | "success" | "error";
 
 export function CartProvider({ children }: { children: React.ReactNode }) {
-  const cart = useShoppingCart();
+  const [cartItems, setCartItems] = useState<CartItemTypes[]>([]);
+  const [status, setStatus] = useState<Status>("idle");
+  const [errorMessage, setErrorMessage] = useState<string[]>([]);
 
-  return <CartContext.Provider value={cart}>{children}</CartContext.Provider>;
+  const updateErrorMessage = useCallback((errorMessage: string) => {
+    setErrorMessage((prev) => [...prev, errorMessage]);
+  }, []);
+
+  useEffect(() => {
+    async function fetchCartItems() {
+      try {
+        setStatus("loading");
+        const cartItemsData = await getShoppingCart();
+        setCartItems(cartItemsData.content);
+        setStatus("success");
+      } catch (e) {
+        setStatus("error");
+        updateErrorMessage(
+          "장바구니 조회 중 오류가 발생했습니다. 잠시 후 다시 시도해 주세요."
+        );
+      }
+    }
+    fetchCartItems();
+  }, [updateErrorMessage]);
+
+  const updateCartItems = useCallback(async () => {
+    try {
+      const cartItemsData = await getShoppingCart();
+      setCartItems(cartItemsData.content);
+    } catch (e) {
+      console.error("장바구니 업데이트 중 오류가 발생했습니다:", e);
+    }
+  }, []);
+
+  const getMatchCartItem = useCallback(
+    (productId: number) => findIsCartItem(cartItems, productId),
+    [cartItems]
+  );
+
+  const checkMax = useCallback(() => isCartFull(cartItems), [cartItems]);
+
+  const contextValue = useMemo(
+    () => ({
+      cartItems,
+      status,
+      errorMessage,
+      updateErrorMessage,
+      updateCartItems,
+      getMatchCartItem,
+      checkMax,
+    }),
+    [
+      cartItems,
+      status,
+      errorMessage,
+      updateErrorMessage,
+      updateCartItems,
+      getMatchCartItem,
+      checkMax,
+    ]
+  );
+  return (
+    <CartContext.Provider value={contextValue}>{children}</CartContext.Provider>
+  );
 }
